src/contoso_chat/requirements.txt
azure-cosmos
azure-identity==1.16.0
azure-search-documents==11.4.0
promptflow==1.10.0
promptflow-tools==1.4.0
promptflow[azure]
python-dotenv==1.0.1
azure-cosmos
jsonlines
promptflow.evals
nbconvert